reapi: enum type prototype for match_ops and test
Problem: we need to have a faux enum type that is shared
within the Go bindings, and a more reasonable test setup
for arbitrary go_<test>.go files.
Solution: add a pkg subdirectory to our go bindings src,
the idea being that we can eventually move more shared
logic between the cli and modules there. In addition,
add an ability to target a single directory with
a go_<test>.go file that (although not build) can be
tested with go test, which is the standard way to do it
with go. We do not (I think) need to worry about having
a binary built for associated go modules (akin to
types.go) because they will ultimately be included
in the reapi_cli or reapi_module that will be
tested. This is just a prototype, meaning that I have
not added the actual enum to the C code, but just
provided the structure of the module that can be
imported into each of the cli and module and
used when that is done.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>