test: adding testing matrix for os release matrix
Problem: we want to make sure that fluxion-go builds across OS releases.
Solution: add each major build to the testing matrix. In addition, we want
to use bookworm-amd64 over bookworm, which does not appear to be updated.
With these newer containers we do not need to build flux-sched to create
the bindings (reapi_cli.so) as they are already installed to /usr

Signed-off-by: vsoch <vsoch@users.noreply.github.com>