ci: add automated release pipeline
Problem: we need to release in synx with flux-sched upstream
Solution: have a release workflow that tests the current master,
and tests and updates the version file and releases, coinciding
with flux-sched. This will likely need further testing and tweaking
as we decide on further customization, but should be a good start.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>