feat: grow support for fluxion
This changeset exposes the (WIP) API to allow a resource
graph to grow, which needs to be a call to the c api
"grow" that gets passed to unpack_at. This is a WIP
because the container build has a custom clone/build of
flux-sched with the grow-api branch, and we will need
to eventually update that.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>