bug: find can return > 0, should not be called an error
This is due to a call to check_array_size
that returns an rc, but it is not actually
a return code, it is the sum of elements
in the array. The function can also return
a negative value (when things go wrong)
but it seems it can look like an error when
it goes right and returns a positive, nonzero
value. For the time being I am tweaking our
reapi error parser to call anything >=0
not an error (nil)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>